pyproject.toml
[project]
name = "course-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "psycopg2-binary>=2.9.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "drf-spectacular>=0.27.0",
    "drf-nested-routers>=0.93.4",
    "python-decouple>=3.8",
    "Pillow>=10.0.0",
    "djangorestframework-simplejwt>=5.3.0",
]

[tool.ruff]
select = ["F", "E", "W", "ARG"]   # include unused rules
ignore = []
exclude = ["migrations"]

[tool.ruff.per-file-ignores]
# Allow unused imports in __init__.py (common in Django)
"__init__.py" = ["F401"]